✨ Add a "silly" rule-based classifier
This doesn't conform to the API in the KNN implementation, but I noticed
that a really simple rule is pretty good at picking out A/B/C/D/E:
whether or not summing each slice is less than 500,000.

The intuition is that if we split the "box region" into five
(44, 56) shape chunks, then the sum of each region is a measure for how
much white there is. "Higher white" => "less likely to be shaded."

From my local testing, shaded in regions tended to have sums around
300,000 whereas unshaded regions tended to be around 550,000.